% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-methods.R
\name{add_confint}
\alias{add_confint}
\alias{add_confint.smooth_estimates}
\title{Add a confidence interval to an existing object}
\usage{
add_confint(object, coverage = 0.95, ...)

\method{add_confint}{smooth_estimates}(object, coverage = 0.95, ...)
}
\arguments{
\item{object}{a R object.}

\item{coverage}{numeric; the coverage for the interval. Must be in the range
0 < \code{coverage} < 1.}

\item{...}{arguments passed to other methods.}
}
\description{
Add a confidence interval to an existing object
}
