% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assemble.R
\name{assemble}
\alias{assemble}
\title{Prepare plots via \code{ggplot2} and assembles them as a list}
\usage{
assemble(object, ...)
}
\arguments{
\item{object}{and R object to plot.}

\item{...}{arguments passed to other methods.}
}
\value{
A list of \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} objects.
}
\description{
Generic function for assembling plot objects created from R objects, using
the \code{ggplot2} package.
}
\author{
Gavin L. Simpson
}
