% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis-dim.R
\name{basis_dim}
\alias{basis_dim}
\alias{basis_dim.mgcv.smooth}
\alias{basis_dim.gam}
\alias{basis_dim.gamm}
\title{Extract basis dimension of a smooth}
\usage{
basis_dim(object, ...)

\method{basis_dim}{mgcv.smooth}(object, ...)

\method{basis_dim}{gam}(object, ...)

\method{basis_dim}{gamm}(object, ...)
}
\arguments{
\item{object}{A fitted GAM(M). Currently \code{\link[mgcv:gam]{mgcv::gam()}} (and anything that
inherits from the \code{"gam"} class, e.g. \code{\link[mgcv:bam]{mgcv::bam()}}) and \code{\link[mgcv:gamm]{mgcv::gamm()}} are
supported.}

\item{...}{Arguments passed to other methods.}
}
\description{
Extract basis dimension of a smooth
}
\examples{
load_mgcv()

df <- data_sim("eg1", n = 200, seed = 1)
m <- bam(y ~ s(x0) + s(x1) + s(x2) + s(x3), data = df)

basis_dim(m)
}
