% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utililties.R
\name{boundary}
\alias{boundary}
\alias{boundary.soap.film}
\alias{boundary.gam}
\title{Extract the boundary of a soap film smooth}
\usage{
boundary(x, ...)

\method{boundary}{soap.film}(x, ...)

\method{boundary}{gam}(x, select, ...)
}
\arguments{
\item{x}{an R object. Currently only objects that inherit from classes
\code{"soap.film"} and \code{"gam"}.}

\item{...}{arguments passed to other methods.}

\item{select}{character; the label of the soap film smooth from which to
extract the boundary.}
}
\value{
A list of lists or data frames specifying the loops that define the
boundary of the soap film smooth.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\seealso{
\link[mgcv:smooth.construct.so.smooth.spec]{mgcv::soap}
}
\keyword{utility}
