% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other-scam-methods.R
\name{coef.scam}
\alias{coef.scam}
\title{Extract coefficients from a fitted \code{scam} model.}
\usage{
\method{coef}{scam}(object, parametrized = TRUE, ...)
}
\arguments{
\item{object}{a model object fitted by \code{scam()}}

\item{parametrized}{logical; extract parametrized coefficients, which
respect the linear inequality constraints of the model.}

\item{...}{other arguments.}
}
\description{
Extract coefficients from a fitted \code{scam} model.
}
