% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utililties.R
\name{data_class}
\alias{data_class}
\title{Vectorized version of \code{data.class}}
\usage{
data_class(df)
}
\arguments{
\item{df}{a data frame or tibble}
}
\description{
Vectorized version of \code{data.class}
}
