% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare-smooths.R
\name{draw.compare_smooths}
\alias{draw.compare_smooths}
\title{Plot comparisons of smooths}
\usage{
\method{draw}{compare_smooths}(object, ncol = NULL, nrow = NULL, guides = "collect", ...)
}
\arguments{
\item{object}{of class \code{"compare_smooths"}, the result of a call to
\code{\link[=compare_smooths]{compare_smooths()}}.}

\item{ncol}{numeric; the numbers of rows and columns over which to
spread the plots}

\item{nrow}{numeric; the numbers of rows and columns over which to
spread the plots}

\item{guides}{character; one of \code{"keep"} (the default), \code{"collect"}, or
\code{"auto"}. Passed to \code{\link[patchwork:plot_layout]{patchwork::plot_layout()}}}

\item{...}{additional arguments passed to \code{\link[patchwork:wrap_plots]{patchwork::wrap_plots()}}.}
}
\description{
Plot comparisons of smooths
}
