% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditional-differences.R
\name{draw.conditional_differences}
\alias{draw.conditional_differences}
\title{Plot conditional differences}
\usage{
\method{draw}{conditional_differences}(
  object,
  facet_scales = "fixed",
  discrete_colour = NULL,
  discrete_fill = NULL,
  xlab = NULL,
  ylab = NULL,
  ...
)
}
\arguments{
\item{object}{an object returned by \code{\link[=conditional_differences]{conditional_differences()}}.}

\item{facet_scales}{character; should facets have the same axis scales
across facets? See \code{\link[ggplot2:facet_wrap]{ggplot2::facet_wrap()}} for details. Options are:
\code{"fixed"} (default), \code{"free_x"}, \code{"free_y"}, and \code{"free"}.}

\item{discrete_colour}{a suitable colour scale to be used when plotting
discrete variables.}

\item{discrete_fill}{a suitable fill scale to be used when plotting
discrete variables.}

\item{xlab}{character; label for the x axis of the plot.}

\item{ylab}{character; label for the y axis of the plot.}

\item{...}{arguments passed to \code{\link[=draw.conditional_values]{draw.conditional_values()}}.}
}
\description{
Display conditional differences with pointwise uncertainty intervals and a
zero reference line. Each comparison is shown in a separate facet; additional
conditioning variables use colour and facets as in \code{\link[=draw.conditional_values]{draw.conditional_values()}}.
}
\examples{
load_mgcv()
df <- data_sim("eg4", seed = 2)
m <- gam(y ~ fac + s(x2, by = fac), data = df, method = "REML")
draw(conditional_differences(m, by = "fac", condition = "x2"))
}
