% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditional-values.R
\name{draw.conditional_values}
\alias{draw.conditional_values}
\title{Plot conditional predictions}
\usage{
\method{draw}{conditional_values}(
  object,
  facet_scales = "fixed",
  discrete_colour = NULL,
  discrete_fill = NULL,
  ...
)
}
\arguments{
\item{object}{an object of class \code{"conditional_values"}, the result of a call
to \code{\link[=conditional_values]{conditional_values()}}.}

\item{facet_scales}{character; should facets have the same axis scales
across facets? See \code{\link[ggplot2:facet_wrap]{ggplot2::facet_wrap()}} for details. Options are:
\code{"fixed"} (default), \code{"free_x"}, \code{"free_y"}, and \code{"free"}.}

\item{discrete_colour}{a suitable colour scale to be used when plotting
discrete variables.}

\item{discrete_fill}{a suitable fill scale to be used when plotting
discrete variables.}

\item{...}{additional arguments passed to \code{\link[patchwork:wrap_plots]{patchwork::wrap_plots()}}.}
}
\description{
Plot conditional predictions
}
