% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw-methods.R
\name{draw.evaluated_2d_smooth}
\alias{draw.evaluated_2d_smooth}
\title{Plot a 2D smooth}
\usage{
\method{draw}{evaluated_2d_smooth}(object, contour = TRUE, xlab, ylab,
  title = NULL, subtitle = NULL, caption = NULL, ...)
}
\arguments{
\item{object}{an object, the result of a call to \code{\link[tsgam:evaluate_smooth]{tsgam::evaluate_smooth()}}.}

\item{contour}{logical; should contours be draw on the plot using
\code{\link[ggplot2:geom_contour]{ggplot2::geom_contour()}}.}

\item{...}{arguments passed to other methods.}
}
\value{
A \code{ggplot2} object
}
\description{
Plots an estimated 2D smooth using ggplot2.
}
\examples{
library("mgcv")
library("ggplot2")
set.seed(2)
dat <- gamSim(1, n = 400, dist = "normal", scale = 2)
m1 <- gam(y ~ s(x0) + s(x1) + s(x2) + s(x3), data = dat, method = "REML")

sm <- evaluate_smooth(m1, "s(x1)")
## draw(sm)

## 2d example
set.seed(2)
dat <- gamSim(2, n = 4000, dist = "normal", scale = 1)
m2 <- gam(y ~ s(x, z, k = 40), data = dat$data, method = "REML")

sm <- evaluate_smooth(m2, "s(x,z)", n = 100)
draw(sm)
}
\author{
Gavin L. Simpson
}
