% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw-methods.R
\name{draw}
\alias{draw}
\title{Generic plotting via \code{ggplot2}}
\usage{
draw(object, ...)
}
\arguments{
\item{object}{and R object to plot.}

\item{...}{arguments passed to other methods.}
}
\value{
A \code{ggplot2} object
}
\description{
Generic plotting via \code{ggplot2}
}
\details{
Generic function for plotting of R objects that uses the \code{ggplot2} package.
}
\author{
Gavin L. Simpson
}
