% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_smooth.R
\name{evaluate_parametric_term}
\alias{evaluate_parametric_term}
\alias{evaluate_parametric_term.gam}
\title{Evaluate parametric model terms}
\usage{
evaluate_parametric_term(object, ...)

\method{evaluate_parametric_term}{gam}(object, term, unconditional = FALSE, ...)
}
\arguments{
\item{object}{an object of class \code{"gam"} or \code{"gamm"}.}

\item{...}{arguments passed to other methods.}

\item{term}{character; which parametric term whose effects are evaluated}

\item{unconditional}{logical; should confidence intervals include the
uncertainty due to smoothness selection? If \code{TRUE}, the corrected Bayesian
covariance matrix will be used.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Returns values of parametric model terms
at values of factor terms and over a grid of covariate values for linear
parametric terms. This function is now deprecated in favour of
\code{\link[=parametric_effects]{parametric_effects()}}.
}
