% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_smooth.R
\name{evaluate_smooth}
\alias{evaluate_smooth}
\title{Evaluate a smooth}
\usage{
evaluate_smooth(object, ...)
}
\arguments{
\item{object}{an object of class \code{"gam"} or \code{"gamm"}.}

\item{...}{arguments passed to other methods.}
}
\value{
A data frame, which is of class \code{"evaluated_1d_smooth"} or
\code{evaluated_2d_smooth}, which inherit from classes \code{"evaluated_smooth"}
and \code{"data.frame"}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Evaluate a smooth at a grid of evenly
spaced value over the range of the covariate associated with the smooth.
Alternatively, a set of points at which the smooth should be evaluated can be
supplied.
}
\details{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} \code{evaluate_smooth()} is deprecated in
favour of \code{\link[=smooth_estimates]{smooth_estimates()}}, which provides a cleaner way to evaluate a
smooth over a range of covariate values. \code{\link[=smooth_estimates]{smooth_estimates()}} can handle a
much wider range of models than \code{evaluate_smooth()} is capable of and
\code{\link[=smooth_estimates]{smooth_estimates()}} is much easier to extend to handle new smooth types.

Most code that uses \code{evaluate_smooth()} should work simply by changing the
function call to \code{\link[=smooth_estimates]{smooth_estimates()}}. However, there are some differences:
\itemize{
\item the \code{newdata} argument becomes \code{data}
}
}
