% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utililties.R
\name{factor_var_names}
\alias{factor_var_names}
\title{Names of any factor variables in model data}
\usage{
factor_var_names(df)
}
\arguments{
\item{df}{a data frame or tibble}
}
\description{
Names of any factor variables in model data
}
