% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family-utils.R
\name{family.gam}
\alias{family.gam}
\alias{family.gamm}
\alias{family.bam}
\title{Family method for GAM objects}
\usage{
\method{family}{gam}(object, ...)

\method{family}{gamm}(object, ...)

\method{family}{bam}(object, ...)
}
\arguments{
\item{object}{a fitted model. Models fitted by \code{\link[mgcv:gam]{mgcv::gam()}}, \code{\link[mgcv:bam]{mgcv::bam()}},
and \code{\link[mgcv:gamm]{mgcv::gamm()}}.}

\item{...}{arguments passed to other methods.}
}
\description{
Provides a \code{\link[stats:family]{stats::family()}} method for a range of GAM objects.
}
