% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family-utils.R
\name{family_name}
\alias{family_name}
\title{Name of family used to fit model}
\usage{
family_name(object, ...)
}
\arguments{
\item{object}{an R object.}

\item{...}{arguments passed to other methods.}
}
\value{
A character vector containing the family name.
}
\description{
Extracts the name of the family used to fit the supplied model.
}
