% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family-utils.R
\name{family_type}
\alias{family_type}
\alias{family_type.family}
\alias{family_type.default}
\title{Extracts the type of family in a consistent way}
\usage{
family_type(object, ...)

\method{family_type}{family}(object, ...)

\method{family_type}{default}(object, ...)
}
\arguments{
\item{object}{an R object. Currently \code{\link[=family]{family()}} objects and anything with a
\code{\link[=family]{family()}} method.}

\item{...}{arguments passed to other methods.}
}
\description{
Extracts the type of family in a consistent way
}
