% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effects.R
\name{fixed_effects}
\alias{fixed_effects}
\alias{fixed_effects.gam}
\alias{fixed_effects.gamm}
\alias{fixed_effects.gamm4}
\alias{fixed_effects.lm}
\alias{fixed_effects.glm}
\alias{fixed_effects.default}
\title{Extract fixed effects estimates from a fitted model}
\usage{
fixed_effects(object, ...)

\method{fixed_effects}{gam}(object, ...)

\method{fixed_effects}{gamm}(object, ...)

\method{fixed_effects}{gamm4}(object, ...)

\method{fixed_effects}{lm}(object, ...)

\method{fixed_effects}{glm}(object, ...)

\method{fixed_effects}{default}(object, ...)
}
\arguments{
\item{object}{a fitted model. Supported classes are models fitted by
\code{\link[mgcv:gam]{mgcv::gam()}}, \code{\link[mgcv:bam]{mgcv::bam()}}, \code{\link[mgcv:gamm]{mgcv::gamm()}}, \code{\link[mgcv:gam]{mgcv::gam()}},
\code{\link[gamm4:gamm4]{gamm4::gamm4()}}, \code{\link[stats:glm]{stats::glm()}}, and \code{\link[stats:lm]{stats::lm()}}.}

\item{...}{arguments passed to other methods}
}
\description{
Extract fixed effects estimates from a fitted model
}
\examples{
load_mgcv()

# run example if lme4 is available
if (require("lme4")) {
  data(sleepstudy, package = "lme4")
  m <- gam(
    Reaction ~ Days + s(Subject, bs = "re") +
      s(Days, Subject, bs = "re"),
    data = sleepstudy, method = "REML"
  )
  fixed_effects(m)
}
}
