% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utililties.R
\name{get_by_smooth}
\alias{get_by_smooth}
\title{Extract an factor-by smooth by name}
\usage{
get_by_smooth(object, term, level)
}
\arguments{
\item{object}{a fitted GAM model object.}

\item{term}{character; the name of a smooth term to extract.}

\item{level}{character; which level of the factor to exrtact the smooth
for.}
}
\value{
A single smooth object, or a list of smooths if several match the
named term.
}
\description{
Extract an factor-by smooth by name
}
