% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utililties.R
\name{get_smooth}
\alias{get_smooth}
\title{Extract an mgcv smooth by name}
\usage{
get_smooth(object, term)
}
\arguments{
\item{object}{a fitted GAM model object.}

\item{term}{character; the name of a smooth term to extract}
}
\value{
A single smooth object, or a list of smooths if several match the
named term.
}
\description{
Extract an mgcv smooth by name
}
