% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utililties.R
\name{get_smooths_by_id}
\alias{get_smooths_by_id}
\alias{get_smooths_by_id.gam}
\alias{get_smooths_by_id.scam}
\alias{get_smooths_by_id.gamm}
\alias{get_smooths_by_id.gamm4}
\alias{get_smooths_by_id.list}
\title{Extract an mgcv smooth given its position in the model object}
\usage{
get_smooths_by_id(object, id)

\method{get_smooths_by_id}{gam}(object, id)

\method{get_smooths_by_id}{scam}(object, id)

\method{get_smooths_by_id}{gamm}(object, id)

\method{get_smooths_by_id}{gamm4}(object, id)

\method{get_smooths_by_id}{list}(object, id)
}
\arguments{
\item{object}{a fitted GAM model object.}

\item{id}{numeric; the position of the smooth in the model object.}
}
\description{
Extract an mgcv smooth given its position in the model object
}
