% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utililties.R
\name{get_vcov}
\alias{get_vcov}
\title{Extract the covariance matrix of the model parameters or a subset of these for
a particular model term}
\usage{
get_vcov(
  object,
  unconditional = FALSE,
  frequentist = FALSE,
  term = NULL,
  by_level = NULL
)
}
\arguments{
\item{object}{an R object from which a covariance matrix shall be extracted.}

\item{unconditional}{logical; if \code{TRUE} (and only if \code{frequentist == FALSE})
then the bayesian smoothing parameter uncertainty-corrected covariance
matrix is returned, if available. Whether it is available depends on which
smoothness selection method was used to fit the model.}

\item{frequentist}{logical; if \code{FALSE}, the default, the bayesian covariance
matrix is returned, otherwise the frequentist covariance matrix.}

\item{term}{character, length 1; return the subset of the covariance matrix
for a particular \strong{smooth} term. Only a single term can be selected.}

\item{by_level}{character; currently unsued.}
}
\value{
a matrix containing the selected covariance matrix.
}
\description{
Extract the covariance matrix of the model parameters or a subset of these for
a particular model term
}
\author{
Gavin L. Simpson
}
