% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gratia-package.R
\docType{package}
\name{gratia-package}
\alias{gratia}
\alias{gratia-package}
\title{gratia: Graceful 'ggplot'-Based Graphics and Other Functions for GAMs Fitted Using 'mgcv'}
\description{
Graceful 'ggplot'-based graphics and utility functions for working with generalized additive models (GAMs) fitted using the 'mgcv' package. Provides a reimplementation of the plot() method for GAMs that 'mgcv' provides, as well as 'tidyverse' compatible representations of estimated smooths.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gavinsimpson.github.io/gratia/}
  \item Report bugs at \url{https://github.com/gavinsimpson/gratia/issues}
}

}
\author{
\strong{Maintainer}: Gavin L. Simpson \email{ucfagls@gmail.com} (\href{https://orcid.org/0000-0002-9084-8413}{ORCID})

Other contributors:
\itemize{
  \item Henrik Singmann (\href{https://orcid.org/0000-0002-4842-3657}{ORCID}) [contributor]
}

}
\keyword{internal}
