% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{gss_vocab}
\alias{gss_vocab}
\title{Data from the General Social Survey (GSS) from the National Opinion Research
Center of the University of Chicago}
\format{
A data frame with 1858 rows and 3 variables:
\itemize{
\item \code{vocab}: numeric; the number of words out of 10 correct on a vocabulary
test.
\item \code{nativeBorn}: factor; Was the respondent born in the US? A factor with
levels \code{no} and \code{yes}.
\item \code{ageGroup}: factor; grouped age of the respondent with levels \code{18-29}
\code{30-39}, \code{40-49}, \code{50-59}, and \verb{60+}.##'
}
}
\description{
A subset of the data from the \code{\link[carData:GSSvocab]{carData::GSSvocab}} dataset from the
\code{carData} package, containing observations from 2016 only.
}
\keyword{data}
