% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-sim.R
\name{gw_f0}
\alias{gw_f0}
\alias{gw_functions}
\alias{gw_f1}
\alias{gw_f2}
\alias{gw_f3}
\title{Gu and Wahba test functions}
\usage{
gw_f0(x, ...)

gw_f1(x, ...)

gw_f2(x, ...)

gw_f3(x, ...)
}
\arguments{
\item{x}{numeric; vector of points to evaluate the function at, on interval
(0,1)}

\item{...}{arguments passed to other methods, ignored.}
}
\description{
Gu and Wahba test functions
}
\examples{
\dontshow{
op <- options(digits = 4)
}
x <- seq(0, 1, length = 6)
gw_f0(x)
gw_f1(x)
gw_f2(x)
gw_f3(x) # should be constant 0
\dontshow{
options(op)
}
}
