% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utililties.R
\name{is_factor_var}
\alias{is_factor_var}
\title{Are variables in a data frame factors?}
\usage{
is_factor_var(df)
}
\arguments{
\item{df}{a data frame or tibble}
}
\description{
Are variables in a data frame factors?
}
