% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utililties.R
\name{is_mgcv_family}
\alias{is_mgcv_family}
\title{Is an object one of mgcv's family objects?}
\usage{
is_mgcv_family(object)
}
\arguments{
\item{object}{the object to test.}
}
\value{
A logical vector of length 1, indicating if \code{object} is one of
\emph{mgcv}'s (\code{TRUE}), or otherwise (\code{FALSE}).
}
\description{
Checks to determine if \code{object} is a family object of one of the three or so
types that \emph{mgcv} produces.
}
