% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utililties.R
\name{is_multivariate_y}
\alias{is_multivariate_y}
\title{Is a model multivariate?}
\usage{
is_multivariate_y(model)
}
\arguments{
\item{model}{a fitted model object; currently only for \code{"gam"} objects}
}
\value{
A logical vector of length 1, indicating if \code{model} is multivariate
(\code{TRUE}), or otherwise (\code{FALSE}).
}
\description{
Determines whether a fitted model (GAM) is truly multivariate or not.
}
