% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utililties.R
\name{load_mgcv}
\alias{load_mgcv}
\title{Load mgcv quietly}
\usage{
load_mgcv()
}
\value{
Returns a logical vectors invisibly, indicating whether the package
was loaded or not.
}
\description{
Simple function that loads the \emph{mgcv} package whilst suppressing the startup
messages that it prints to the console.
}
