% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lp-matrix.R
\name{lp_matrix}
\alias{lp_matrix}
\alias{lp_matrix.gam}
\title{Return the linear prediction matrix}
\usage{
lp_matrix(model, ...)

\method{lp_matrix}{gam}(model, data = NULL, ...)
}
\arguments{
\item{model}{a fitted model}

\item{...}{arguments passed to other methods and \code{predict} methods including
\link[mgcv:predict.gam]{mgcv::predict.gam} and \link[mgcv:predict.bam]{mgcv::predict.bam}}

\item{data}{a data frame of values at which to return the linear prediction
matrix.}
}
\description{
Return the linear prediction matrix
}
\examples{
load_mgcv()
\dontshow{op <- options(digits = 3, cli.unicode = FALSE)}
df <- data_sim("eg1", seed = 1)
m <- gam(y ~ s(x0) + s(x1) + s(x2) + s(x3), data = df)

# linear prediction matrix for observed data
xp <- lp_matrix(m)
xp

# the object `xp` *is* a matrix
class(xp)
# but we print like a tibble to avoid spamming the R console

# linear predictor matrix for new data set
ds <- data_slice(m, x2 = evenly(x2))
xp <- lp_matrix(m, data = ds)
xp
\dontshow{options(op)}
}
