% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributional-models.R
\name{lss_parameters}
\alias{lss_parameters}
\title{General names of LSS parameters for each GAM family}
\usage{
lss_parameters(object)
}
\description{
General names of LSS parameters for each GAM family
}
\keyword{internal}
