% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth-concurvity.R
\name{model_concurvity}
\alias{model_concurvity}
\alias{concrvity}
\title{Concurvity of an estimated GAM}
\usage{
model_concurvity(
  model,
  terms = everything(),
  type = c("all", "estimate", "observed", "worst"),
  pairwise = FALSE
)

concrvity(
  model,
  terms = everything(),
  type = c("all", "estimate", "observed", "worst"),
  pairwise = FALSE
)
}
\description{
Concurvity of an estimated GAM
}
\examples{
## simulate data with concurvity...
set.seed(8)
n <- 200
df <- tibble(t = sort(runif(n)),
             x = gratia::f2_fun(t) + rnorm(n) * 3,
             y = sin(4 * pi * t) + exp(x / 20) + rnorm(n) * 0.3)

## fit model
m <- gam(y ~ s(t, k = 15) + s(x, k = 15), data = df, method = "REML")

## overall concurvity
o_conc <- concrvity(m)
draw(o_conc)

## pairwise concurvity
p_conc <- concrvity(m)
draw(p_conc)
}
