% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utililties.R
\name{n_eta}
\alias{n_eta}
\alias{n_eta.gam}
\title{The Number of linear predictors in model}
\usage{
n_eta(model, ...)

\method{n_eta}{gam}(model, ...)
}
\arguments{
\item{model}{a fitted model. Currently, only models inheriting from class
\code{"gam"} are supported.}

\item{...}{arguments passed to methods.}
}
\value{
An integer vector of length 1 containing the number of linear
predictors in the model.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}} Extracts the number of linear predictors
from the fitted model.
}
\keyword{utility}
