% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utililties.R
\name{n_smooths}
\alias{n_smooths}
\alias{n_smooths.default}
\alias{n_smooths.gam}
\alias{n_smooths.gamm}
\title{How many smooths in a fitted model}
\usage{
n_smooths(object)

\method{n_smooths}{default}(object)

\method{n_smooths}{gam}(object)

\method{n_smooths}{gamm}(object)
}
\arguments{
\item{object}{a fitted GAM or related model. Typically the result of a call
to \code{\link[mgcv:gam]{mgcv::gam()}}, \code{\link[mgcv:bam]{mgcv::bam()}}, or \code{\link[mgcv:gamm]{mgcv::gamm()}}.}
}
\description{
How many smooths in a fitted model
}
