% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utililties.R
\name{null_deviance}
\alias{null_deviance}
\alias{null_deviance.default}
\title{Extract the null deviance of a fitted model}
\usage{
null_deviance(model, ...)

\method{null_deviance}{default}(model, ...)
}
\arguments{
\item{model}{a fitted model}

\item{...}{arguments passed to other methods}
}
\description{
Extract the null deviance of a fitted model
}
