% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnose.R
\name{observed_fitted_plot}
\alias{observed_fitted_plot}
\title{Plot of fitted against observed response values}
\usage{
observed_fitted_plot(model, ylab = NULL, xlab = NULL, title = NULL,
  subtitle = NULL, caption = NULL)
}
\arguments{
\item{model}{a fitted model. Currently only class \code{"gam"}.}

\item{ylab}{character or expression; the label for the y axis. If not
supplied, a suitable label will be generated.}

\item{xlab}{character or expression; the label for the y axis. If not
supplied, a suitable label will be generated.}

\item{title}{character or expression; the title for the plot. See
\code{\link[ggplot2:labs]{ggplot2::labs()}}.}

\item{subtitle}{character or expression; the subtitle for the plot. See
\code{\link[ggplot2:labs]{ggplot2::labs()}}.}

\item{caption}{character or expression; the plot caption. See
\code{\link[ggplot2:labs]{ggplot2::labs()}}.}
}
\description{
Plot of fitted against observed response values
}
