% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overview.R
\name{overview}
\alias{overview}
\alias{overview.gam}
\title{Provides an overview of a model and the terms in that model}
\usage{
overview(model, ...)

\method{overview}{gam}(
  model,
  parametric = TRUE,
  random_effects = TRUE,
  dispersion = NULL,
  freq = FALSE,
  ...
)
}
\arguments{
\item{model}{a fitted model object to overview.}

\item{...}{arguments passed to other methods.}
}
\description{
Provides an overview of a model and the terms in that model
}
