% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utililties.R
\name{parametric_terms}
\alias{parametric_terms}
\alias{parametric_terms.default}
\alias{parametric_terms.gam}
\title{Names of any parametric terms in a GAM}
\usage{
parametric_terms(model, ...)

\method{parametric_terms}{default}(model, ...)

\method{parametric_terms}{gam}(model, ...)
}
\arguments{
\item{model}{a fitted model.}

\item{...}{arguments passed to other methods.}
}
\description{
Names of any parametric terms in a GAM
}
