% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitted_values.R
\name{post_link_funs}
\alias{post_link_funs}
\title{A list of transformation functions named for LSS parameters in a GAMLSS}
\usage{
post_link_funs(
  location = identity_fun,
  scale = identity_fun,
  shape = identity_fun,
  skewness = identity_fun,
  kurtosis = identity_fun,
  power = identity_fun,
  pi = identity_fun
)
}
\description{
A list of transformation functions named for LSS parameters in a GAMLSS
}
\keyword{internal}
