% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{ref_sims}
\alias{ref_sims}
\title{Reference simulation data}
\format{
A named list of simulated data sets created by \code{\link[=data_sim]{data_sim()}}.
}
\description{
A set of reference objects for testing \code{\link[=data_sim]{data_sim()}}.
}
\keyword{data}
