% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw-utilities.R
\name{reorder_fs_smooth_terms}
\alias{reorder_fs_smooth_terms}
\title{Reorder random factor smooth terms to place factor last}
\usage{
reorder_fs_smooth_terms(smooth)
}
\arguments{
\item{smooth}{an mgcv smooth object}
}
\description{
Reorder random factor smooth terms to place factor last
}
\keyword{internal}
