% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw-utilities.R
\name{reorder_tensor_smooth_terms}
\alias{reorder_tensor_smooth_terms}
\title{Reorder tensor product terms for nicer plotting}
\usage{
reorder_tensor_smooth_terms(smooth)
}
\arguments{
\item{smooth}{an mgcv smooth object}
}
\description{
If a tensor product smooth of 3 or more terms contains a 2d marginal smooth,
we will get nicer output from \code{\link[=smooth_estimates]{smooth_estimates()}} and hence a nicer plot
from the \code{\link[=draw.smooth_estimates]{draw.smooth_estimates()}} method if we reorder the terms of the
smooth such that we vary the terms in the 2d marginal first, and any other
terms vary more slowly when we generate data to evaluate the smooth at. This
results in automatically generated data that focuses on the (or the first if
more than one) 2d marginal smooth, with the end result that
\code{\link[=smooth_estimates]{smooth_estimates()}} shows how that 2d smooth changes with the other terms
involved in the smooth.
}
\keyword{internal}
