% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rootogram.R
\name{rootogram}
\alias{rootogram}
\alias{rootogram.gam}
\title{Computes data used to draw a rootogram diagnostic plot}
\usage{
rootogram(object, ...)

\method{rootogram}{gam}(object, max_count = NULL, breaks = "Sturges", ...)
}
\arguments{
\item{object}{an R object}

\item{...}{arguments passed to other methods}

\item{max_count}{integer; the largest count to consider}

\item{breaks}{for continuous responses, how to group the response. Can be
anything that is acceptable as the \code{breaks} argument of
\code{\link[graphics:hist]{graphics::hist.default()}}}
}
\description{
Computes data used to draw a rootogram diagnostic plot
}
\examples{
load_mgcv()
df <- data_sim("eg1", n = 1000, dist = "poisson", scale = 0.1, seed = 6)

# A poisson example
m <- gam(y ~ s(x0, bs = "cr") + s(x1, bs = "cr") + s(x2, bs = "cr") +
         s(x3, bs = "cr"), family = nb, data = df, method = "REML")
rg <- rootogram(m)
rg
draw(rg) # plot the rootogram

# A Gaussian example
df <- data_sim("eg1", dist = "normal", seed = 2)
m <- gam(y ~ s(x0) + s(x1) + s(x2) + s(x3), data = df, method = "REML")
draw(rootogram(m, breaks = "FD"), type = "suspended")
}
