% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot-functions.R
\name{scale_fill_partial_effect}
\alias{scale_fill_partial_effect}
\title{Default diverging red-blue colour palette for partial effects}
\usage{
scale_fill_partial_effect(
  name = "Partial effect",
  ...,
  na.value = "grey50",
  guide = "colourbar",
  direction = -1
)
}
\arguments{
\item{name}{The name of the scale. Used as the legend title. If \code{NULL} the
legend title will be omitted.}

\item{...}{arguments passed to \code{\link[ggplot2:continuous_scale]{ggplot2::continuous_scale()}}.}

\item{na.value}{Missing values will be replaced with this value (colour).}

\item{guide}{A function used to create a guide or its name. See
\code{\link[ggplot2:guides]{ggplot2::guides()}} for more information.}

\item{direction}{Sets the order of colours in the scale. If \code{1}, the
default, colours are as output by \code{\link[RColorBrewer:ColorBrewer]{RColorBrewer::brewer.pal()}}. If \code{-1},
the order of colours is reversed.}
}
\description{
Default diverging red-blue colour palette for partial effects
}
