% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utililties.R
\name{seq_min_max}
\alias{seq_min_max}
\title{Create a sequence of evenly-spaced values}
\usage{
seq_min_max(x, n)
}
\arguments{
\item{x}{numeric; vector over which evenly-spaced values are returned}

\item{n}{numeric; the number of evenly-spaced values to return}
}
\value{
A numeric vector of length \code{n}.
}
\description{
Creates a sequence of \code{n} evenly-spaced values over the range
\code{min(x)} -- \code{max(x)}.
}
\examples{
\dontshow{set.seed(1)}
x <- rnorm(10)
n <- 10L
seq_min_max(x, n = n)
}
