% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utililties.R
\name{shift_values}
\alias{shift_values}
\title{Shift numeric values in a data frame by an amount \code{eps}}
\usage{
shift_values(df, h, i, FUN = "+")
}
\arguments{
\item{df}{a data frame or tibble.}

\item{h}{numeric; the amount to shift values in \code{df} by.}

\item{i}{logical; a vector indexing columns of \code{df} that should not be
included in the shift.}

\item{FUN}{function; a function to applut the shift. Typically \code{+} or \code{-}.}
}
\description{
Shift numeric values in a data frame by an amount \code{eps}
}
