% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{smallAges}
\alias{smallAges}
\title{Lead-210 age-depth measurements for Small Water}
\format{
A data frame with 12 rows and 7 variables.
}
\source{
Simpson, G.L. (Unpublished data).
}
\description{
A dataset containing lead-210 based age depth measurements for the SMALL1
core from Small Water.
}
\details{
The variables are as follows:
\itemize{
\item \code{Depth}
\item \code{Drymass}
\item \code{Date}
\item \code{Age}
\item \code{Error}
\item \code{SedAccRate}
\item \code{SedPerCentChange}
}
}
\keyword{data}
