% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth-coefs.R
\name{smooth_coef_indices}
\alias{smooth_coef_indices}
\title{Indices of the parametric terms for a particular smooth}
\usage{
smooth_coef_indices(smooth)
}
\arguments{
\item{smooth}{an object that inherits from class \code{mgcv.smooth}}
}
\value{
A numeric vector of indices.
}
\description{
Returns a vector of indices of the parametric terms that represent the
supplied smooth. Useful for extracting model coefficients and columns
of their covariance matrix.
}
\seealso{
\code{\link[=smooth_coefs]{smooth_coefs()}} for extracting the coefficients for a particular
smooth.
}
\author{
Gavin L. Simpson
}
