% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth-data.R
\name{smooth_data}
\alias{smooth_data}
\title{Generate regular data over the covariates of a smooth}
\usage{
smooth_data(
  model,
  id,
  n = 100,
  n_2d = NULL,
  n_3d = NULL,
  n_4d = NULL,
  offset = NULL,
  include_all = FALSE,
  var_order = NULL,
  envir = NULL
)
}
\arguments{
\item{model}{a fitted model}

\item{id}{the number ID of the smooth within \code{model} to process.}

\item{n}{numeric; the number of new observations to generate.}

\item{n_2d}{numeric; the number of new observations to generate for the
second dimension of a 2D smooth. \emph{Currently ignored}.}

\item{n_3d}{numeric; the number of new observations to generate for the third
dimension of a 3D smooth.}

\item{n_4d}{numeric; the number of new observations to generate for the
dimensions higher than 2 (!) of a \emph{k}D smooth (\emph{k} >= 4). For example, if
the smooth is a 4D smooth, each of dimensions 3 and 4 will get \code{n_4d}
new observations.}

\item{offset}{numeric; value of the model offset to use.}

\item{include_all}{logical; include all covariates involved in the smooth?
if \code{FALSE}, only the covariates involved in the smooth will be included in
the returned data frame. If \code{TRUE}, a representative value will be included
for all other covariates in the model that aren't actually used in the
smooth. This can be useful if you want to pass the returned data frame on
to \code{\link[mgcv:PredictMat]{mgcv::PredictMat()}}.}

\item{var_order}{character; the order in which the terms in the smooth
should be processed. Only useful for tensor products with at least one
2d marginal smooth.}

\item{envir}{an optional environment supplying functions and constants used
in model expressions. The available model formula environment is used when
\code{NULL}. Covariate observations should be supplied in \code{data}.}
}
\description{
Generate regular data over the covariates of a smooth
}
\examples{
\dontshow{
op <- options(cli.unicode = FALSE, pillar.sigfig = 4)
}
load_mgcv()
df <- data_sim("eg1", seed = 42)
m <- bam(y ~ s(x0) + s(x1) + s(x2) + s(x3), data = df)

# generate data over range of x1 for smooth s(x1)
smooth_data(m, id = 2)

# generate data over range of x1 for smooth s(x1), with typical value for
# other covariates in the model
smooth_data(m, id = 2, include_all = TRUE)

\dontshow{
options(op)
}
}
