% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utililties.R
\name{smooth_dim}
\alias{smooth_dim}
\alias{smooth_dim.gam}
\alias{smooth_dim.gamm}
\alias{smooth_dim.mgcv.smooth}
\title{Dimension of a smooth}
\usage{
smooth_dim(object)

\method{smooth_dim}{gam}(object)

\method{smooth_dim}{gamm}(object)

\method{smooth_dim}{mgcv.smooth}(object)
}
\arguments{
\item{object}{an R object. See Details for list of supported objects.}
}
\value{
A numeric vector of dimensions for each smooth.
}
\description{
Extracts the dimension of an estimated smooth.
}
\details{
This is a generic function with methods for objects of class
\code{"gam"}, \code{"gamm"}, and \code{"mgcv.smooth"}.
}
\author{
Gavin L. Simpson
}
