% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utililties.R
\name{smooth_terms}
\alias{smooth_terms}
\title{List the variables involved in smooths}
\usage{
smooth_terms(object, ...)
}
\arguments{
\item{object}{an R object the result of a call to \code{\link[mgcv:gam]{mgcv::gam()}},
\code{\link[mgcv:bam]{mgcv::bam()}}, or \code{\link[mgcv:gamm]{mgcv::gamm()}}, or that inherits from classes \code{"gam"} or
\code{"mgcv.smooth"}, or \code{"fs.interaction"}.}

\item{...}{arguments passed to other methods. Currently unused.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
