% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/too_far.R
\name{to_na}
\alias{to_na}
\title{Sets the elements of vector to \code{NA}}
\usage{
to_na(x, i)
}
\arguments{
\item{x}{vector of values}

\item{i}{vector of values used to subset \code{x}}
}
\value{
Returns \code{x} with possibly some elements set to \code{NA}
}
\description{
Given a vector \code{i} indexing the elements of \code{x}, sets the selected elements
of \code{x} to \code{NA}.
}
