% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/too_far.R
\name{too_far_to_na}
\alias{too_far_to_na}
\title{Set rows of data to \code{NA} if the lie too far from a reference set of values}
\usage{
too_far_to_na(smooth, input, reference, cols, dist = NULL)
}
\arguments{
\item{smooth}{an mgcv smooth object}

\item{input}{data frame containing the input observations and the columns to
be set to \code{NA}}

\item{reference}{data frame containing the reference values}

\item{cols}{character vector of columns whose elements will be set to \code{NA} if
the data lies too far from the reference set}

\item{dist}{numeric, the distance from the reference set beyond which
elements of \code{input} will be set to \code{NA}}
}
\description{
Set rows of data to \code{NA} if the lie too far from a reference set of values
}
