% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utililties.R
\name{transform_fun}
\alias{transform_fun}
\title{Transform estimated values and confidence intervals by applying a function}
\usage{
transform_fun(object, fun = NULL, ...)
}
\arguments{
\item{object}{an object to apply the transform function to.}

\item{fun}{the function to apply.}

\item{...}{additional arguments passed to methods.}
}
\description{
Transform estimated values and confidence intervals by applying a function
}
