% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-slice.R
\name{typical_values}
\alias{typical_values}
\alias{typical_values.gam}
\alias{typical_values.data.frame}
\title{Typical values of model covariates}
\usage{
typical_values(object, ...)

\method{typical_values}{gam}(object, vars = everything(), data = NULL, ...)

\method{typical_values}{data.frame}(object, vars = everything(), ...)
}
\arguments{
\item{object}{a fitted GAM(M) model.}

\item{...}{arguments passed to other methods.}

\item{vars}{terms to include or exclude from the returned object. Uses
tidyselect principles.}

\item{data}{an optional data frame supplying covariate classes. By default,
these are taken from the stored model frame; typical values come from
the fitted model's covariate summaries.}
}
\description{
Typical values of model covariates
}
