% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplers.R
\name{user_draws}
\alias{user_draws}
\title{Handle user-supplied posterior draws}
\usage{
user_draws(model, draws, ...)
}
\arguments{
\item{model}{a fitted R model. Currently only models fitted by \code{mgcv::gam()}
or \code{mgcv::bam()}, or return an object that \emph{inherits} from such objects are
supported. Here, "inherits" is used in a loose fashion; models fitted by
\code{scam::scam()} are support even though those models don't strictly inherit
from class \code{"gam"} as far as \code{inherits()} is concerned.}

\item{draws}{matrix; user supplied posterior draws to be used when
\code{method = "user"}.}

\item{...}{arguments passed to methods.}
}
\description{
Handle user-supplied posterior draws
}
