% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utililties.R
\name{vars_from_label}
\alias{vars_from_label}
\title{Returns names of variables from a smooth label}
\usage{
vars_from_label(label)
}
\arguments{
\item{label}{character; a length 1 character vector containing the label of
a smooth.}
}
\description{
Returns names of variables from a smooth label
}
\examples{

vars_from_label("s(x1)")
vars_from_label("t2(x1,x2,x3)")
}
