% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utililties.R
\name{which_smooth}
\alias{which_smooth}
\title{Identify a smooth term by it's label}
\usage{
which_smooth(object, term)
}
\arguments{
\item{object}{a fitted GAM.}

\item{term}{character; the (partial) term label with which to identify the required smooth.}
}
\description{
Identify a smooth term by it's label
}
