% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utililties.R
\name{which_smooths}
\alias{which_smooths}
\alias{which_smooths.default}
\alias{which_smooths.gam}
\alias{which_smooths.bam}
\alias{which_smooths.gamm}
\title{Identify a smooth term by it's label}
\usage{
which_smooths(object, ...)

\method{which_smooths}{default}(object, ...)

\method{which_smooths}{gam}(object, terms, ...)

\method{which_smooths}{bam}(object, terms, ...)

\method{which_smooths}{gamm}(object, terms, ...)
}
\arguments{
\item{object}{a fitted GAM.}

\item{...}{arguments passed to other methods.}

\item{terms}{character; one or more (partial) term labels with which to identify
required smooths.}
}
\description{
Identify a smooth term by it's label
}
