% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{zooplankton}
\alias{zooplankton}
\title{Madison lakes zooplankton data}
\format{
A data frame
}
\source{
Pedersen EJ, Miller DL, Simpson GL, Ross N. 2018. Hierarchical
generalized additive models: an introduction with mgcv. \emph{PeerJ Preprints}
\strong{6}:e27320v1 \doi{10.7287/peerj.preprints.27320v1}.
}
\description{
The Madison lake zooplankton data are from a long-term study in seasonal
dynamics of zooplankton, collected by the Richard Lathrop. The data were
collected from a chain of lakes in Wisconsin (Mendota, Monona, Kegnonsa,
and Waubesa) approximately bi-weekly from 1976 to 1994. They consist of
samples of the zooplankton communities, taken from the deepest point of each
lake via vertical tow. The data are provided by the Wisconsin Department of
Natural Resources and their collection and processing are fully described in
Lathrop (2000).
}
\details{
Each record consists of counts of a given zooplankton taxon taken from a
subsample from a single vertical net tow, which was then scaled to account
for the relative volume of subsample versus the whole net sample and the area
of the net tow and rounded to the nearest 1000 to give estimated population
density per m2 for each taxon at each point in time in each sampled lake.
}
\references{
Lathrop RC. (2000). Madison Wisonsin Lakes Zooplankton 1976--1994.
Environmental Data Initiative.
}
\keyword{data}
