% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utililties.R
\name{add_constant}
\alias{add_constant}
\title{Add a constant to estimated values}
\usage{
add_constant(object, constant = NULL, ...)
}
\arguments{
\item{object}{a object to add a constant to.}

\item{constant}{the constant to add.}

\item{...}{additional arguments passed to methods.}
}
\description{
Add a constant to estimated values
}
