% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-methods.R
\name{add_constant}
\alias{add_constant}
\alias{add_constant.evaluated_smooth}
\alias{add_constant.smooth_estimates}
\alias{add_constant.mgcv_smooth}
\alias{add_constant.parametric_effects}
\alias{add_constant.tbl_df}
\alias{add_constant.evaluated_parametric_term}
\title{Add a constant to estimated values}
\usage{
add_constant(object, constant = NULL, ...)

\method{add_constant}{evaluated_smooth}(object, constant = NULL, ...)

\method{add_constant}{smooth_estimates}(object, constant = NULL, ...)

\method{add_constant}{mgcv_smooth}(object, constant = NULL, ...)

\method{add_constant}{parametric_effects}(object, constant = NULL, ...)

\method{add_constant}{tbl_df}(object, constant = NULL, column = NULL, ...)

\method{add_constant}{evaluated_parametric_term}(object, constant = NULL, ...)
}
\arguments{
\item{object}{a object to add a constant to.}

\item{constant}{the constant to add.}

\item{...}{additional arguments passed to methods.}

\item{column}{character; for the \code{"tbl_df"} method, which column to add the
constant too.}
}
\value{
Returns \code{object} but with the estimate shifted by the addition of
the supplied constant.
}
\description{
Add a constant to estimated values
}
\author{
Gavin L. Simpson
}
