% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-slice.R
\name{data_combos}
\alias{data_combos}
\alias{data_combos.gam}
\title{All combinations of factor levels plus typical values of continuous variables}
\usage{
data_combos(object, ...)

\method{data_combos}{gam}(object, vars = everything(), complete = TRUE, ...)
}
\arguments{
\item{object}{a fitted model object.}

\item{...}{arguments passed to methods.}

\item{vars}{terms to include or exclude from the returned object. Uses
tidyselect principles.}

\item{complete}{logical; should all combinations of factor levels be
returned? If \code{FALSE}, only those combinations of levels observed in the
model are retained.}
}
\description{
All combinations of factor levels plus typical values of continuous variables
}
