% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-sim.R
\name{data_sim}
\alias{data_sim}
\title{Simulate example data for fitting GAMs}
\usage{
data_sim(
  model = "eg1",
  n = 400,
  scale = 2,
  theta = 3,
  dist = c("normal", "poisson", "binary", "negbin", "tweedie", "gamma", "ocat",
    "ordered categorical"),
  n_cat = 4,
  cuts = c(-1, 0, 5),
  seed = NULL
)
}
\arguments{
\item{model}{character; either \code{"egX"} where \code{X} is an integer \code{1:7}, or
the name of a model. See Details for possible options.}

\item{n}{numeric; the number of observations to simulate.}

\item{scale}{numeric; the level of noise to use.}

\item{theta}{numeric; the dispersion parameter \eqn{\theta} to use. The
default is entirely arbitrary, chosen only to provide simulated data that
exhibits extra dispersion beyond that assumed by under a Poisson.}

\item{dist}{character; a sampling distribution for the response
variable.}

\item{n_cat}{integer; the number of categories for categorical response.
Currently only used for \code{distr \%in\% c("ocat", "ordered categorical")}.}

\item{cuts}{numeric; vector of cut points on the latent variable, excluding
the end points \code{-Inf} and \code{Inf}. Must be one fewer than the number of
categories: \code{length(cuts) == n_cat - 1}.}

\item{seed}{numeric; the seed for the random number generator. Passed to
\code{\link[base:Random]{base::set.seed()}}.}
}
\description{
A tidy reimplementation of the functions implemented in \code{\link[mgcv:gamSim]{mgcv::gamSim()}}
that can be used to fit GAMs. An new feature is that the sampling
distribution can be applied to all the example types.
}
\examples{
\dontshow{op <- options(pillar.sigfig = 5, cli.unicode = FALSE)
}
data_sim("eg1", n = 100, seed = 1)

# an ordered categorical response
data_sim("eg1", n = 100, dist = "ocat", n_cat = 4, cuts = c(-1, 0, 5))
\dontshow{options(op)}
}
