% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-slice.R
\name{data_slice}
\alias{data_slice}
\title{Prepare a data slice through covariates}
\usage{
data_slice(object, ...)
}
\arguments{
\item{object}{an R model object.}

\item{...}{arguments passed to other methods.}
}
\description{
Prepare a data slice through covariates
}
