% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derivatives.R
\name{derivatives}
\alias{derivatives}
\alias{derivatives.default}
\alias{derivatives.gamm}
\alias{derivatives.gam}
\title{Derivatives of estimated smooths via finite differences}
\usage{
derivatives(object, ...)

\method{derivatives}{default}(object, ...)

\method{derivatives}{gamm}(object, ...)

\method{derivatives}{gam}(object, term, newdata, order = 1L,
  type = c("forward", "backward", "central"), n = 200, eps = 1e-07,
  unconditional = FALSE, frequentist = FALSE, offset = NULL, ...)
}
\arguments{
\item{object}{an R object to compute derivatives for.}

\item{...}{arguments passed to other methods.}

\item{term}{character; vector of one or more smooth terms for which
derivatives are required. If missing, derivatives for all smooth terms
will be returned. Can be a partial match to a smooth term.}

\item{newdata}{a data frame containing the values of the model covariates
at which to evaluate the first derivatives of the smooths.}

\item{order}{numeric; the order of derivative.}

\item{type}{character; the type of finite difference used. One of
\code{"forward"}, \code{"backward"}, or \code{"central"}.}

\item{n}{numeric; the number of points to evaluate the derivative at.}

\item{eps}{numeric; the finite difference.}

\item{unconditional}{logical; use smoothness selection-corrected Bayesian
covariance matrix?}

\item{frequentist}{logical; use the frequentist covariance matrix?}

\item{offset}{numeric; a value to use for any offset term}
}
\description{
Derivatives of estimated smooths via finite differences
}
\examples{

library("mgcv")
set.seed(42)
dat <- gamSim(1, n = 400, dist = "normal", scale = 2, verbose = FALSE)
mod <- gam(y ~ s(x0) + s(x1) + s(x2) + s(x3), data = dat, method = "REML")

## first derivative of all smooths
derivatives(mod)

## second derivative of smooth of x2 only using central finite difference
derivatives(mod, "s(x2)", order = 2, type = "central")
}
\author{
Gavin L. Simpson
}
