% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw-methods.R
\name{draw.basis}
\alias{draw.basis}
\title{Plot basis functions}
\usage{
\method{draw}{basis}(
  object,
  legend = FALSE,
  use_facets = TRUE,
  labeller = NULL,
  ylab = "Value",
  title = NULL,
  subtitle = NULL,
  caption = NULL,
  ncol = NULL,
  nrow = NULL,
  angle = NULL,
  guides = "keep",
  ...
)
}
\arguments{
\item{object}{an object, the result of a call to \code{\link[=basis]{basis()}}.}

\item{legend}{logical; should a legend by drawn to indicate basis functions?}

\item{use_facets}{logical; for factor by smooths, use facets to show the
basis functions for each level of the factor? If \code{FALSE}, a separate ggplot
object will be created for each level and combined using
\code{\link[patchwork:wrap_plots]{patchwork::wrap_plots()}}. \strong{Currently ignored}.}

\item{labeller}{a labeller function with which to label facets. The default
is to use \code{\link[ggplot2:labellers]{ggplot2::label_both()}}.}

\item{ylab}{character or expression; the label for the y axis. If not
supplied, a suitable label will be generated from \code{object}.}

\item{title}{character or expression; the title for the plot. See
\code{\link[ggplot2:labs]{ggplot2::labs()}}.}

\item{subtitle}{character or expression; the subtitle for the plot. See
\code{\link[ggplot2:labs]{ggplot2::labs()}}.}

\item{caption}{character or expression; the plot caption. See
\code{\link[ggplot2:labs]{ggplot2::labs()}}.}

\item{ncol, nrow}{numeric; the numbers of rows and columns over which to
spread the plots}

\item{angle}{numeric; the angle at which the x axis tick labels are to be
drawn passed to the \code{angle} argument of \code{\link[ggplot2:guide_axis]{ggplot2::guide_axis()}}.}

\item{guides}{character; one of \code{"keep"} (the default), \code{"collect"}, or
\code{"auto"}. Passed to \code{\link[patchwork:plot_layout]{patchwork::plot_layout()}}}

\item{...}{arguments passed to other methods. Not used by this method.}
}
\value{
A \code{patchwork} object.
}
\description{
Plots basis functions using ggplot2
}
\examples{
load_mgcv()
df <- data_sim("eg1", n = 400, seed = 42)
m <- gam(y ~ s(x0) + s(x1) + s(x2) + s(x3), data = df, method = "REML")

bf <- basis(m)
draw(bf)

bf <- basis(m, "s(x2)")
draw(bf)
}
\author{
Gavin L. Simpson
}
