% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rootogram.R
\name{draw.rootogram}
\alias{draw.rootogram}
\title{Draw the computed rootogram}
\usage{
\method{draw}{rootogram}(
  object,
  type = c("hanging", "standing", "suspended"),
  sqrt = TRUE,
  ref_line = TRUE,
  warn_limits = TRUE,
  ylab = NULL,
  xlab = NULL,
  ...
)
}
\arguments{
\item{object}{and R object to plot.}

\item{type}{character}

\item{sqrt}{logical}

\item{ref_line}{logical}

\item{warn_limits}{logical}

\item{xlab, ylab}{character; labels for the x and y axis of the rootogram.
May be missing (\code{NULL}), in which case suitable labels will be used.}

\item{...}{arguments passed to other methods.}
}
\value{
A 'ggplot' object.
}
\description{
Draw the computed rootogram
}
