% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_smooth2.R
\name{eval_smooth}
\alias{eval_smooth}
\title{S3 methods to evaulate individual smooths}
\usage{
eval_smooth(smooth, ...)
}
\arguments{
\item{smooth}{currently an object that inherits from class \code{mgcv_smooth}}

\item{...}{arguments assed to other methods}
}
\description{
S3 methods to evaulate individual smooths
}
