% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_smooth.R
\name{evaluate_smooth}
\alias{evaluate_smooth}
\title{Evaluate a smooth}
\usage{
evaluate_smooth(object, smooth, n = 200, newdata = NULL,
  unconditional = FALSE, inc.mean = FALSE)
}
\arguments{
\item{object}{an object of class \code{"gam"} or \code{"gamm"}.}

\item{smooth}{character; a single smooth to evaluate.}

\item{n}{numeric; the number of points over the range of the covariate at which to evaluate the smooth.}

\item{newdata}{a vector or data frame of points at which to evaluate the smooth.}

\item{unconditional}{logical; should confidence intervals include the uncertainty due to smoothness selection? If \code{TRUE}, the corrected Bayesian covariance matrix will be used.}

\item{inc.mean}{logical; should the uncertainty in the model constant term be included in the standard error of the evaluate values of the smooth? Currently not implemented.}
}
\description{
Evaluate a smooth at a grid of evenly spaced value over the range of the covariate associated with the smooth. Alternatively, a set of points at which the smooth should be evaluated can be supplied.
}
\examples{
library("mgcv")
set.seed(2)
dat <- gamSim(1, n = 400, dist = "normal", scale = 2)
mod <- gam(y ~ s(x0) + s(x1) + s(x2) + s(x3), data = dat, method = "REML")

head(evaluate_smooth(mod, "s(x1)"))
}
