% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior-samples.R
\name{fitted_samples}
\alias{fitted_samples}
\title{Draw fitted values from the posterior distribution}
\usage{
fitted_samples(model, ...)
}
\arguments{
\item{model}{a fitted model of the supported types}

\item{...}{arguments passed to other methods}
}
\description{
Draw fitted values from the posterior distribution
}
