% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utililties.R
\name{get_smooths_by_id}
\alias{get_smooths_by_id}
\title{Extract an mgcv smooth given its position in the model object}
\usage{
get_smooths_by_id(object, id)
}
\arguments{
\item{object}{a fitted GAM model object.}

\item{id}{numeric; the position of the smooth in the model object.}
}
\description{
Extract an mgcv smooth given its position in the model object
}
