% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utililties.R
\name{is_by_smooth}
\alias{is_by_smooth}
\alias{is_factor_by_smooth}
\alias{is_continuous_by_smooth}
\title{Tests for by variable smooths}
\usage{
is_by_smooth(smooth)

is_factor_by_smooth(smooth)

is_continuous_by_smooth(smooth)
}
\arguments{
\item{smooth}{an object of class \code{"mgcv.smooth"}}
}
\value{
A logical vector.
}
\description{
Functions to check if a smooth is a by-variable one and to test of the type
of by-variable smooth is a factor-smooth or a continous-smooth interaction.
}
\author{
Gavin L. Simpson
}
