% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utililties.R
\name{is_factor_var}
\alias{is_factor_var}
\alias{is_numeric_var}
\title{Vectorised checks for variable types}
\usage{
is_factor_var(df)

is_numeric_var(df)
}
\arguments{
\item{df}{a data frame or tibble}
}
\description{
Vectorised checks for variable types
}
