% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utililties.R
\name{is_mgcv_smooth}
\alias{is_mgcv_smooth}
\title{Check if objects are smooths or are a particular type of smooth}
\usage{
is_mgcv_smooth(smooth)
}
\arguments{
\item{smooth}{an R object, typically a list}
}
\description{
Check if objects are smooths or are a particular type of smooth
}
