% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic-plots.R
\name{observed_fitted_plot}
\alias{observed_fitted_plot}
\title{Plot of fitted against observed response values}
\usage{
observed_fitted_plot(
  model,
  ylab = NULL,
  xlab = NULL,
  title = NULL,
  subtitle = NULL,
  caption = NULL,
  point_col = "grey20",
  point_alpha = 1
)
}
\arguments{
\item{model}{a fitted model. Currently only class \code{"gam"}.}

\item{ylab}{character or expression; the label for the y axis. If not
supplied, a suitable label will be generated.}

\item{xlab}{character or expression; the label for the y axis. If not
supplied, a suitable label will be generated.}

\item{title}{character or expression; the title for the plot. See
\code{\link[ggplot2:labs]{ggplot2::labs()}}.}

\item{subtitle}{character or expression; the subtitle for the plot. See
\code{\link[ggplot2:labs]{ggplot2::labs()}}.}

\item{caption}{character or expression; the plot caption. See
\code{\link[ggplot2:labs]{ggplot2::labs()}}.}

\item{point_col}{colour used to draw points in the plots. See
\code{\link[graphics:par]{graphics::par()}} section \strong{Color Specification}. This is passed to
the individual plotting functions, and therefore affects the points of
all plots.}

\item{point_alpha}{numeric; alpha transparency for points in plots.}
}
\description{
Plot of fitted against observed response values
}
