% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parametric-effects.R
\name{parametric_effects}
\alias{parametric_effects}
\title{Estimated values for parametric model terms}
\usage{
parametric_effects(object, ...)
}
\arguments{
\item{object}{a fitted model object.}

\item{...}{arguments passed to other methods.}

\item{term}{character; which parametric term whose effects are evaluated.}

\item{ci_level}{numeric; the coverage required for the confidence interval.
Currently ignored.}
}
\description{
Estimated values for parametric model terms
}
