% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penalty.R
\name{penalty}
\alias{penalty}
\alias{penalty.gam}
\alias{penalty.mgcv.smooth}
\alias{penalty.tensor.smooth}
\alias{penalty.t2.smooth}
\title{Extract and tidy penalty matrices}
\usage{
penalty(object, ...)

\method{penalty}{gam}(object, smooth = NULL, rescale = FALSE, ...)

\method{penalty}{mgcv.smooth}(object, rescale = FALSE, ...)

\method{penalty}{tensor.smooth}(object, margins = FALSE, ...)

\method{penalty}{t2.smooth}(object, margins = FALSE, ...)
}
\arguments{
\item{object}{a fitted GAM or a smooth.}

\item{...}{additional arguments passed to methods.}

\item{smooth}{character; vector of smooths to extract penalty matrices for.
If \code{NULL}, penalty matrices for all smooths in \code{object} are extracted.}

\item{rescale}{logical; by default, \emph{mgcv} will scale the penalty matrix for
better performance in \code{\link[mgcv:gamm]{mgcv::gamm()}}. If \code{rescale} is \code{TRUE}, this scaling
will be undone to put the penalty matrix back on the original scale.}

\item{margins}{logical; extract the penalty matrices for the tensor
product or the margainl smooths of the tensor product?}
}
\description{
Extract and tidy penalty matrices
}
\examples{
load_mgcv()
dat <- data_sim("eg4", n = 400, seed = 42)
m <- gam(y ~ s(x0) + s(x1) + s(x2, by = fac),
         data = dat, method = "REML")
penalty(m)

# for a specific smooth
penalty(m, smooth = "s(x2):fac1")
}
\author{
Gavin L. Simpson
}
