% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior-samples.R
\name{posterior_samples}
\alias{posterior_samples}
\title{Draw samples from the posterior distribution of an estimated model}
\usage{
posterior_samples(model, ...)
}
\arguments{
\item{model}{a fitted model of the supported types}

\item{...}{arguments passed to other methods}

\item{newdata}{data frame; new observations at which the posterior draws
from the model should be evaluated. If not supplied, the data used to fit
the model will be used for \code{newdata}, if available in \code{model}.}

\item{n}{numeric; the number of posterior samples to return.}

\item{seed}{numeric; a random seed for the simulations.}

\item{scale}{character;}

\item{freq}{logical; \code{TRUE} to return the frequentist covariance matrix of
the parameter estimators, \code{FALSE} to return the Bayesian posterior
covariance matrix of the parameters.}

\item{unconditional}{logical; if \code{TRUE} (and \code{freq == FALSE}) then the
Bayesian smoothing parameter uncertainty corrected covariance matrix is
returned, if available.}

\item{weights}{numeric; a vector of prior weights. If \code{newdata} is null
then defaults to \code{object[["prior.weights"]]}, otherwise a vector of ones.}
}
\description{
Draw samples from the posterior distribution of an estimated model
}
