% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.R
\name{quantile_residuals}
\alias{quantile_residuals}
\alias{quantile_residuals.gam}
\alias{quantile_residuals.glm}
\title{Randomised residuals}
\usage{
quantile_residuals(model, type = c("pit", "quantile"), seed = NULL, ...)

\method{quantile_residuals}{gam}(model, type = c("pit", "quantile"), seed = NULL, ...)

\method{quantile_residuals}{glm}(model, type = c("pit", "quantile"), seed = NULL, ...)
}
\arguments{
\item{model}{a fitted model object.}

\item{type}{character; which type of randomised residual to return}

\item{seed}{integer; the random seed to use when generating randomised
residuals. Can be missing, in which case the current state residuals are
computed using the current state of the random number generator.}

\item{...}{arguments passed to other methods.}
}
\description{
Randomised residuals
}
