% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd-helpers.R
\name{rtw}
\alias{rtw}
\title{Simulator for tweedie LSS models}
\usage{
rtw(mu, p, phi)
}
\arguments{
\item{mu}{numeric vector of mean values of Tweedie distribution.}

\item{p}{numeric vector of values for the power parameter of the Tweedie
distribution.}

\item{phi}{numeric vector of values for the scale parameter \eqn{\phi}{phi}
of the Tweedie distribution.}
}
\description{
Simulate random deviates from a Tweedie distribution with given parameters
\eqn{\mu}{mu}, \eqn{p}{p}, and \eqn{\phi}{phi}. Works with vector values for
all parameters, unlike the version on \emph{mgcv}.
}
