% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth-coefs.R
\name{smooth_coefs}
\alias{smooth_coefs}
\alias{smooth_coefs.gam}
\alias{smooth_coefs.bam}
\alias{smooth_coefs.gamm}
\alias{smooth_coefs.gamm4}
\alias{smooth_coefs.list}
\alias{smooth_coefs.mgcv.smooth}
\alias{smooth_coefs.scam}
\title{Coefficients for a particular smooth}
\usage{
smooth_coefs(object, ...)

\method{smooth_coefs}{gam}(object, term, ...)

\method{smooth_coefs}{bam}(object, term, ...)

\method{smooth_coefs}{gamm}(object, term, ...)

\method{smooth_coefs}{gamm4}(object, term, ...)

\method{smooth_coefs}{list}(object, term, ...)

\method{smooth_coefs}{mgcv.smooth}(object, model, ...)

\method{smooth_coefs}{scam}(object, model, ...)
}
\arguments{
\item{object}{a fitted GAM(M) object, or, for the \code{"mgcv.smooth"} method,
an object that inherits from class \code{mgcv.smooth}.}

\item{...}{arguments passed to other methods.}

\item{term}{character; the label of the smooth whose coefficients will be
returned.}

\item{model}{a fitted GAM(M) object.}
}
\value{
A numeric vector of model coefficients.
}
\description{
Returns a vector of model coefficients of the parametric terms that represent
the supplied smooth.
}
\examples{

load_mgcv()
df <- data_sim("eg1", seed = 2)
m <- gam(y ~ s(x0) + s(x1) + s(x2) + s(x3), data = df, method = "REML")
smooth_coefs(m, term = "s(x2)")
}
\seealso{
\code{\link[=smooth_coef_indices]{smooth_coef_indices()}} for extracting the indices of the
coefficients for a particular smooth.
}
\author{
Gavin L. Simpson
}
