% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utililties.R
\name{smooth_coefs}
\alias{smooth_coefs}
\title{Indices of the parametric terms for a particular smooth}
\usage{
smooth_coefs(smooth)
}
\arguments{
\item{smooth}{an object that inherits from class \code{mgcv.smooth}}
}
\value{
A numeric vector of indices.
}
\description{
Returns a vector of indices of the parametric terms that represent the
supplied smooth. Useful for extracting model coefficients and columns
of their covariance matrix.
}
\author{
Gavin L. Simpson
}
