% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior-samples.R
\name{smooth_samples}
\alias{smooth_samples}
\alias{smooth_samples.gam}
\title{Draw predicted values from the posterior distribution}
\usage{
smooth_samples(model, ...)

\method{smooth_samples}{gam}(model, term = NULL, n = 1,
  newdata = NULL, seed = NULL, freq = FALSE, unconditional = FALSE,
  weights = NULL, ncores = 1L, n_vals = 200, ...)
}
\arguments{
\item{model}{a fitted model of the supported types}

\item{...}{arguments passed to other methods}

\item{term}{character; select which smooth's posterior to draw from.
The default (\code{NULL}) means the posteriors of all smooths in \code{model}
wil be sampled from. If supplied, a character vector of requested terms.}

\item{n}{numeric; the number of posterior samples to return.}

\item{newdata}{data frame; new observations at which the posterior draws
from the model should be evaluated. If not supplied, the data used to fit
the model will be used for \code{newdata}, if available in \code{model}.}

\item{seed}{numeric; a random seed for the simulations.}

\item{freq}{logical; \code{TRUE} to return the frequentist covariance matrix of
the parameter estimators, \code{FALSE} to return the Bayesian posterior
covariance matrix of the parameters.}

\item{unconditional}{logical; if \code{TRUE} (and \code{freq == FALSE}) then the
Bayesian smoothing parameter uncertainty corrected covariance matrix is
returned, if available.}

\item{weights}{numeric; a vector of prior weights. If \code{newdata} is null
then defaults to \code{object[["prior.weights"]]}, otherwise a vector of ones.}

\item{ncores}{number of cores for generating random variables from a
multivariate normal distribution. Passed to \code{\link[mvnfast:rmvn]{mvnfast::rmvn()}}.
Parallelization will take place only if OpenMP is supported (but appears
to work on Windows with current \code{R}).}

\item{n_vals}{numeric; how many locations to evaluate the smooth at if
\code{newdata} not supplied}
}
\value{
A tibble (data frame) with 3 columns containing the posterior
predicted values in long format. The columns are
\itemize{
\item \code{row} (integet) the row of \code{newdata} that each posterior draw relates to,
\item \code{draw} (integer) an index, in range \code{1:n}, indicating which draw each row
relates to,
\item \code{response} (numeric) the predicted response for the indicated row of
\code{newdata}.
}
}
\description{
Predicted values of the response drawn from the posterior distribution of
fitted model, created via \code{simulate()} (e.g. \code{\link[=simulate.gam]{simulate.gam()}})
and returned in a tidy, long, format.
}
\examples{
suppressPackageStartupMessages(library("mgcv"))
\dontshow{set.seed(2)}
dat <- gamSim(1, n = 400, dist = "normal", scale = 2)
m1 <- gam(y ~ s(x0) + s(x1) + s(x2) + s(x3), data = dat, method = "REML")

## smooth_samples(m1, term = "s(x0)", n = 5, seed = 42)
}
\author{
Gavin L. Simpson

Gavin L. Simpson
}
