% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis-funs.R
\name{tidy_basis}
\alias{tidy_basis}
\title{A tidy basis representation of a smooth object}
\usage{
tidy_basis(smooth, data, at = NULL)
}
\arguments{
\item{smooth}{a smooth object.}

\item{data}{a data frame containing the variables used in \code{smooth}.}

\item{at}{a data frame containing values of the smooth covariate(s) at which
the basis should be evaluated.}
}
\value{
A tibble.
}
\description{
Takes an object of class \code{mgcv.smooth} and returns a tidy representation
of the basis.
}
\author{
Gavin L. Simpson
}
