% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utililties.R
\name{which_smooths}
\alias{which_smooths}
\title{Identify a smooth term by it's label}
\usage{
which_smooths(object, terms)
}
\arguments{
\item{object}{a fitted GAM.}

\item{terms}{character; one or more (partial) term labels with which to identify
required smooths.}
}
\description{
Identify a smooth term by it's label
}
